refactor: reorganize test structure into __tests__ directories
- Move all 47 test files from scattered locations to organized __tests__ directories
- Create consistent test structure: [directory]/__tests__/[test-file].test.[ts|tsx]
- Update Jest configuration with proper test patterns and coverage exclusions
- Update GitHub Actions workflow to reflect new test file paths
- Add comprehensive testing documentation in docs/testing-structure.md

Benefits:
- Clean separation between source code and tests
- Better organization with tests grouped by feature proximity
- Improved maintainability with tests located near source code
- Enhanced CI/CD with better test discovery and parallel execution
- Consistent structure across the entire project

Closes: Test structure reorganization for better maintainability
