+++ .github/workflows/run-tests.yml
@@ -33,21 +33,21 @@ jobs:
           - test-group: 1
             is-first-group: true
             test-files: >-
-              lib/data-fetching.test.ts
-              lib/utils.test.ts
+              lib/__tests__/data-fetching.test.ts
+              lib/__tests__/utils.test.ts
               lib/__tests__/template-constants.test.ts
               lib/__tests__/storage-service.test.ts
               lib/__tests__/path-utils.test.ts
-              lib/constants.test.ts
-              lib/supabase-server.test.ts
-              lib/stripe-server.test.ts
-              hooks/use-command-menu.test.tsx
-              hooks/use-search.test.ts
-              components/house-filters.test.tsx
-              components/wohnung-overview-modal.test.tsx
-              components/search-error-boundary.test.tsx
-              components/search-result-item.test.tsx
-              components/finance-edit-modal.test.tsx
+              lib/__tests__/constants.test.ts
+              lib/__tests__/supabase-server.test.ts
+              lib/__tests__/stripe-server.test.ts
+              hooks/__tests__/use-command-menu.test.tsx
+              hooks/__tests__/use-search.test.ts
+              components/__tests__/house-filters.test.tsx
+              components/__tests__/wohnung-overview-modal.test.tsx
+              components/__tests__/search-error-boundary.test.tsx
+              components/__tests__/search-result-item.test.tsx
+              components/__tests__/finance-edit-modal.test.tsx
           - test-group: 2
             is-first-group: false
             test-files: >-
@@ -56,16 +56,16 @@ jobs:
               hooks/__tests__/use-cloud-storage-navigation.test.ts
               hooks/__tests__/use-cloud-storage-store.test.ts
               hooks/__tests__/use-modal-store-timeout.test.ts
-              components/ui/custom-combobox.test.tsx
-              components/ui/video-player.test.tsx
-              components/ui/label-with-tooltip.test.tsx
-              components/ui/button-with-tooltip.test.tsx
-              app/(dashboard)/wohnungen/client.test.tsx
-              app/(dashboard)/todos/client-wrapper.test.tsx
-              app/(dashboard)/betriebskosten/client-wrapper.test.tsx
-              app/(dashboard)/betriebskosten/client-wrapper-layout.test.tsx
-              app/wohnungen-actions.test.ts
-              app/subscription-locked/page.test.tsx
+              components/ui/__tests__/custom-combobox.test.tsx
+              components/ui/__tests__/video-player.test.tsx
+              components/ui/__tests__/label-with-tooltip.test.tsx
+              components/ui/__tests__/button-with-tooltip.test.tsx
+              app/(dashboard)/wohnungen/__tests__/client.test.tsx
+              app/(dashboard)/todos/__tests__/client-wrapper.test.tsx
+              app/(dashboard)/betriebskosten/__tests__/client-wrapper.test.tsx
+              app/(dashboard)/betriebskosten/__tests__/client-wrapper-layout.test.tsx
+              app/__tests__/wohnungen-actions.test.ts
+              app/subscription-locked/__tests__/page.test.tsx
           - test-group: 3
             is-first-group: false
             test-files: >-
@@ -81,7 +81,7 @@ jobs:
               lib/__tests__/mention-suggestion-popup.test.ts
               lib/__tests__/path-utils-virtual-folders.test.ts
               lib/__tests__/directory-cache.test.ts
-              integration/landing-page-flow.test.tsx
+              integration/__tests__/landing-page-flow.test.tsx
     
     # Job name is set by the matrix strategy
