+++ app/(dashboard)/finanzen/__tests__/client-wrapper.test.tsx
