+++ app/api/apartments/[apartmentId]/tenant/[tenantId]/details/__tests__/route.test.ts
