+++ app/api/finanzen/charts/__tests__/route.test.ts
