+++ components/__tests__/search-result-group.test.tsx
