+++ components/__tests__/wohnung-overview-modal.test.tsx
@@ -1,22 +1,15 @@
-import React from 'react';
-import { render, screen, fireEvent, waitFor, act } from '@testing-library/react';
-import { WohnungOverviewModal } from '../wohnung-overview-modal';
+import { render, screen, act } from '@testing-library/react';
+import { WohnungOverviewModal } from './wohnung-overview-modal';
 import { useModalStore } from '@/hooks/use-modal-store';
-import { toast } from '@/hooks/use-toast';
 
 // Mock timers
 jest.useFakeTimers();
 
-// Mock the hooks
+// Mock the modal store
 jest.mock('@/hooks/use-modal-store');
-jest.mock('@/hooks/use-toast');
+const mockUseModalStore = useModalStore as jest.MockedFunction<typeof useModalStore>;
 
-// Mock server actions
-jest.mock('@/app/mieter-actions', () => ({
-  deleteTenantAction: jest.fn(),
-}));
-
-// Mock format utilities
+// Mock the format utilities
 jest.mock('@/utils/format', () => ({
   formatNumber: (num: number, fractionDigits: number = 2) => {
     return new Intl.NumberFormat('de-DE', {
@@ -30,19 +23,23 @@ jest.mock('@/utils/format', () => ({
   }).format(num)} €`,
 }));
 
-// Mock fetch
-global.fetch = jest.fn();
+// Mock the toast hook
+jest.mock('@/hooks/use-toast', () => ({
+  toast: jest.fn(),
+}));
 
-const mockUseModalStore = useModalStore as jest.MockedFunction<typeof useModalStore>;
-const mockToast = toast as jest.MockedFunction<typeof toast>;
+// Mock server actions
+jest.mock('@/app/mieter-actions', () => ({
+  deleteTenantAction: jest.fn(),
+}));
 
 describe('WohnungOverviewModal', () => {
   const mockWohnungData = {
     id: '1',
-    name: 'Wohnung 1A',
-    groesse: 75,
+    name: 'Wohnung 1',
+    hausName: 'Test Haus',
+    groesse: 80,
     miete: 1200,
-    hausName: 'Musterstraße 1',
     mieter: [
       {
         id: '1',
@@ -57,31 +54,28 @@ describe('WohnungOverviewModal', () => {
         id: '2',
         name: 'Anna Schmidt',
         email: 'anna@example.com',
-        telefon: null,
+        telefon: '+49987654321',
         einzug: '2022-01-01',
         auszug: '2022-12-31',
-        status: 'moved_out' as const,
+        status: 'inactive' as const,
       },
     ],
   };
 
-  const defaultMockStore = {
-    isWohnungOverviewModalOpen: false,
-    wohnungOverviewData: undefined,
-    wohnungOverviewLoading: false,
-    wohnungOverviewError: undefined,
-    closeWohnungOverviewModal: jest.fn(),
-    setWohnungOverviewLoading: jest.fn(),
-    setWohnungOverviewError: jest.fn(),
-    setWohnungOverviewData: jest.fn(),
-    refreshWohnungOverviewData: jest.fn(),
-    openTenantModal: jest.fn(),
-  };
-
   beforeEach(() => {
-    jest.clearAllMocks();
     jest.clearAllTimers();
-    mockUseModalStore.mockReturnValue(defaultMockStore as any);
+    mockUseModalStore.mockReturnValue({
+      isWohnungOverviewModalOpen: false,
+      wohnungOverviewData: null,
+      wohnungOverviewLoading: false,
+      wohnungOverviewError: undefined,
+      closeWohnungOverviewModal: jest.fn(),
+      setWohnungOverviewLoading: jest.fn(),
+      setWohnungOverviewError: jest.fn(),
+      setWohnungOverviewData: jest.fn(),
+      refreshWohnungOverviewData: jest.fn(),
+      openTenantModal: jest.fn(),
+    } as any);
   });
 
   afterEach(() => {
@@ -91,206 +85,99 @@ describe('WohnungOverviewModal', () => {
     jest.clearAllTimers();
   });
 
-  it('renders nothing when modal is closed', () => {
+  it('should not render when modal is closed', () => {
     render(<WohnungOverviewModal />);
     expect(screen.queryByText('Wohnungs-Übersicht')).not.toBeInTheDocument();
   });
 
-  it('displays loading state', () => {
+  it('should render summary cards with correct data when modal is open', () => {
     mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
       isWohnungOverviewModalOpen: true,
+      wohnungOverviewData: mockWohnungData,
+      wohnungOverviewLoading: false,
+      wohnungOverviewError: undefined,
+      closeWohnungOverviewModal: jest.fn(),
+      setWohnungOverviewLoading: jest.fn(),
+      setWohnungOverviewError: jest.fn(),
+      setWohnungOverviewData: jest.fn(),
+      refreshWohnungOverviewData: jest.fn(),
+      openTenantModal: jest.fn(),
+    } as any);
+
+    render(<WohnungOverviewModal />);
+    
+    expect(screen.getByText('Wohnungs-Übersicht: Wohnung 1')).toBeInTheDocument();
+    expect(screen.getByText('Haus: Test Haus')).toBeInTheDocument();
+    expect(screen.getByText('80,00 m²')).toBeInTheDocument();
+    expect(screen.getByText('1.200,00 €')).toBeInTheDocument();
+  });
+
+  it('should show loading skeleton when loading', () => {
+    mockUseModalStore.mockReturnValue({
+      isWohnungOverviewModalOpen: true,
+      wohnungOverviewData: null,
       wohnungOverviewLoading: true,
+      wohnungOverviewError: undefined,
+      closeWohnungOverviewModal: jest.fn(),
+      setWohnungOverviewLoading: jest.fn(),
+      setWohnungOverviewError: jest.fn(),
+      setWohnungOverviewData: jest.fn(),
+      refreshWohnungOverviewData: jest.fn(),
+      openTenantModal: jest.fn(),
     } as any);
 
     render(<WohnungOverviewModal />);
     
     expect(screen.getByText('Wohnungs-Übersicht')).toBeInTheDocument();
-    // Check for skeleton elements
+    // Check for skeleton elements by class name
     const skeletonElements = document.querySelectorAll('.animate-pulse');
     expect(skeletonElements.length).toBeGreaterThan(0);
   });
 
-  it('displays error state with retry button', () => {
-    const mockRetry = jest.fn();
-    mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
-      isWohnungOverviewModalOpen: true,
-      wohnungOverviewError: 'Test error message',
-      setWohnungOverviewLoading: mockRetry,
-    } as any);
-
-    render(<WohnungOverviewModal />);
-    
-    expect(screen.getByText('Fehler beim Laden')).toBeInTheDocument();
-    expect(screen.getByText('Test error message')).toBeInTheDocument();
-    expect(screen.getByText('Erneut versuchen')).toBeInTheDocument();
-  });
-
-  it('displays empty state when no mieter exist', () => {
+  it('should show empty state with summary cards when no tenants', () => {
     const emptyWohnungData = {
       ...mockWohnungData,
       mieter: [],
     };
 
     mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
       isWohnungOverviewModalOpen: true,
       wohnungOverviewData: emptyWohnungData,
+      wohnungOverviewLoading: false,
+      wohnungOverviewError: undefined,
+      closeWohnungOverviewModal: jest.fn(),
+      setWohnungOverviewLoading: jest.fn(),
+      setWohnungOverviewError: jest.fn(),
+      setWohnungOverviewData: jest.fn(),
+      refreshWohnungOverviewData: jest.fn(),
+      openTenantModal: jest.fn(),
     } as any);
 
     render(<WohnungOverviewModal />);
     
     expect(screen.getByText('Keine Mieter')).toBeInTheDocument();
     expect(screen.getByText('Diese Wohnung hat noch keine Mieter.')).toBeInTheDocument();
+    // Summary cards should still be visible
+    expect(screen.getByText('80,00 m²')).toBeInTheDocument();
   });
 
-  it('displays wohnung data and mieter list', () => {
+  it('should calculate price per square meter correctly', () => {
     mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
       isWohnungOverviewModalOpen: true,
       wohnungOverviewData: mockWohnungData,
+      wohnungOverviewLoading: false,
+      wohnungOverviewError: undefined,
+      closeWohnungOverviewModal: jest.fn(),
+      setWohnungOverviewLoading: jest.fn(),
+      setWohnungOverviewError: jest.fn(),
+      setWohnungOverviewData: jest.fn(),
+      refreshWohnungOverviewData: jest.fn(),
+      openTenantModal: jest.fn(),
     } as any);
 
     render(<WohnungOverviewModal />);
     
-    expect(screen.getByText('Wohnungs-Übersicht: Wohnung 1A')).toBeInTheDocument();
-    expect(screen.getByText('Haus: Musterstraße 1')).toBeInTheDocument();
-    expect(screen.getByText('Max Mustermann')).toBeInTheDocument();
-    expect(screen.getByText('Anna Schmidt')).toBeInTheDocument();
-  });
-
-  it('handles edit mieter action', () => {
-    const mockOpenTenantModal = jest.fn();
-    mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
-      isWohnungOverviewModalOpen: true,
-      wohnungOverviewData: mockWohnungData,
-      openTenantModal: mockOpenTenantModal,
-    } as any);
-
-    render(<WohnungOverviewModal />);
-    
-    const editButtons = screen.getAllByTitle('Mieter bearbeiten');
-    fireEvent.click(editButtons[0]);
-    
-    expect(mockOpenTenantModal).toHaveBeenCalled();
-  });
-
-  it('handles view mieter details action', () => {
-    mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
-      isWohnungOverviewModalOpen: true,
-      wohnungOverviewData: mockWohnungData,
-    } as any);
-
-    render(<WohnungOverviewModal />);
-    
-    const viewButtons = screen.getAllByTitle('Details anzeigen');
-    fireEvent.click(viewButtons[0]);
-    
-    expect(mockToast).toHaveBeenCalledWith({
-      title: "Mieter Details",
-      description: 'Detailansicht für "Max Mustermann" wird geöffnet.',
-      variant: "default",
-    });
-  });
-
-  it('handles contact mieter action with email', () => {
-    mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
-      isWohnungOverviewModalOpen: true,
-      wohnungOverviewData: mockWohnungData,
-    } as any);
-
-    render(<WohnungOverviewModal />);
-    
-    const contactButtons = screen.getAllByTitle(/E-Mail an/);
-    fireEvent.click(contactButtons[0]);
-    
-    // The main behavior we want to test is that the toast is called
-    expect(mockToast).toHaveBeenCalledWith({
-      title: "E-Mail wird geöffnet",
-      description: "E-Mail an Max Mustermann wird in Ihrem Standard-E-Mail-Programm geöffnet.",
-      variant: "default",
-    });
-  });
-
-  it('handles contact mieter action with no contact data', () => {
-    const mieterWithoutContact = {
-      ...mockWohnungData,
-      mieter: [{
-        id: '3',
-        name: 'Test User',
-        email: '',
-        telefon: '',
-        einzug: '2023-01-01',
-        auszug: null,
-        status: 'active' as const,
-      }],
-    };
-
-    mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
-      isWohnungOverviewModalOpen: true,
-      wohnungOverviewData: mieterWithoutContact,
-    } as any);
-
-    render(<WohnungOverviewModal />);
-    
-    const contactButtons = screen.getAllByTitle('Keine Kontaktdaten verfügbar');
-    expect(contactButtons[0]).toBeDisabled();
-  });
-
-  it('closes modal when dialog is closed', () => {
-    const mockClose = jest.fn();
-    mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
-      isWohnungOverviewModalOpen: true,
-      wohnungOverviewData: mockWohnungData,
-      closeWohnungOverviewModal: mockClose,
-    } as any);
-
-    render(<WohnungOverviewModal />);
-    
-    // Simulate dialog close by pressing escape
-    fireEvent.keyDown(document, { key: 'Escape' });
-    
-    // The close function should be called when the dialog's onOpenChange is triggered
-    // This is handled by the Dialog component internally
-    expect(screen.getByText('Wohnungs-Übersicht: Wohnung 1A')).toBeInTheDocument();
-  });
-
-  it('formats dates correctly', () => {
-    mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
-      isWohnungOverviewModalOpen: true,
-      wohnungOverviewData: mockWohnungData,
-    } as any);
-
-    render(<WohnungOverviewModal />);
-    
-    // Check if dates are formatted correctly (German format)
-    // The dates might be formatted differently, so let's check for the presence of date elements
-    const dateElements = screen.getAllByText(/\d{1,2}\.\d{1,2}\.\d{4}/);
-    expect(dateElements.length).toBeGreaterThan(0);
-  });
-
-  it('displays contact links correctly', () => {
-    mockUseModalStore.mockReturnValue({
-      ...defaultMockStore,
-      isWohnungOverviewModalOpen: true,
-      wohnungOverviewData: mockWohnungData,
-    } as any);
-
-    render(<WohnungOverviewModal />);
-    
-    // Check email link
-    const emailLink = screen.getByText('max@example.com');
-    expect(emailLink).toHaveAttribute('href', 'mailto:max@example.com');
-    
-    // Check phone link
-    const phoneLink = screen.getByText('+49123456789');
-    expect(phoneLink).toHaveAttribute('href', 'tel:+49123456789');
+    // Price per sqm should be 1200 / 80 = 15 €/m²
+    expect(screen.getByText('15,00 €/m²')).toBeInTheDocument();
   });
 });
